requirements.txt
beautifulsoup4==4.9.3
certifi==2021.10.8
chardet==3.0.4
idna==2.8
lxml==4.7.1
Pillow==8.4.0
pytesseract==0.3.6
requests==2.27.0
selenium==4.7.2
soupsieve==2.3.1
urllib3==1.26
